Modified StoreProduct and DBStoreProduct to use long for ItemID.
Modified StoreProductQuery interface to use a different method name when
getting Store Products using an ItemID.

Implemented DBStoreProductQuery to add, get, remove, update StoreProducts.
Implemented DBStoreProductQueryTest to use JUnit for testing of
DBStoreProductQuery.

+++ src/main/java/edu/cpp/cs580/Database/Objects/Interfaces/StoreProduct.java
@@ -1,16 +1,15 @@
 package edu.cpp.cs580.Database.Objects.Interfaces;
 
-import java.math.BigInteger;
-import java.util.Date;
+import java.sql.Date;
 
 public interface StoreProduct {
-	public BigInteger getItemID();
+	public long getItemID();
 	public double getPrice();
 	public Date getPriceDate();
 	public int getStoreID();
 	public String getStoreProductID();
 	public String getURL();
-	public void setItemID(BigInteger id);
+	public void setItemID(long id);
 	public void setPrice(double pr);
 	public void setPriceDate(Date date);
 	public void setStoreID(int id);
